requirements.txt
requests
pytest
allure-pytest
pyyaml
cryptography
python-dotenv